src/kwork_parser_bot/kwork_parser_bot/requirements.txt
aiogram==3.0.0b7
redis
python-dotenv
pydantic
loguru
jinja2
stringcase
apscheduler
cron-validator
sqlalchemy==2.0.17
asyncpg
psycopg2
fakeredis
black
pytest==7.4.0
pytest-asyncio
aiohttp
websockets
redis-om
aiohttp_socks